+++ CHANGELOG.md
@@ -25,6 +25,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 
 ### Removed
 - Remove Kotlin data-class generated methods from API.
+- Remove all deprecated API.
 
 ## [2.2.0] - 2018-10-31
 [2.2.0]: https://bitbucket.org/atlassian/virtual-users/branches/compare/release-2.2.0%0Drelease-2.1.5