+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://bitbucket.org/atlassian/virtual-users/branches/compare/master%0Drelease-2.2.0
+[Unreleased]: https://bitbucket.org/atlassian/virtual-users/branches/compare/master%0Drelease-3.0.0
+
+## [3.0.0] - 2018-11-06
+[3.0.0]: https://bitbucket.org/atlassian/virtual-users/branches/compare/release-3.0.0%0Drelease-2.2.0
 
 ### Removed
 - Remove Kotlin data-class generated methods from API.